Massive change to documentation sourcing
We used to ignore enums and had multiple files that we had to exclude do to quirks. We now have a quirks file to handle the quirks, allowing those files in. Likewise, we now treat enums and structures very differently than how we treat functions. This should allow for easier parsing and more sane results. I also added a file to hold onto unique wow data type declarations to get rid of a lot of warnings.